Guess 1 in Semaphore::down()
When we get woken, we are most likely to be the first waiter who was woken by
the Semaphore::up() call, not by another Semaphore::down() call. So we don't
expect to see the need_to_wake_bit set, as it's cleared by the Semaphore::up()
call.

This is also true in practice: seeing 1 after waking is most probable, and is
about 10 times more probable then seeing 1 | need_to_wake_bit.